requirements.txt
ultralytics
opencv-python
numpy
matplotlib
streamlit
openai
python-dotenv
requests
Pillow